MFC r359057:
Remove spurious warning about invalid VPD data.

The warning used to be displayed for valid VPDs about 512B or above in
size.  Fix the size check and add a break while here so that the routine
stops if if detects any problem.

Tested with "pciconf -lV"

Reviewed by:	kib@, jhb@
Sponsored by:	Chelsio Communications
Differential Revision:	https://reviews.freebsd.org/D23679


git-svn-id: https://svn.freebsd.org/base/stable/12@359313 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f