MFC r339880 (arichardson): Fix get_maxfds() in jevents
If RLIM_INFINITY == -1ULL (such as on macOS) the min() call will result
in a value of less than 1 being returned. This causes nftw() to fail
with EINVAL.

While touching this file also fix includes to work on Linux/macOS and don't
declare snprintf since it may have different attributes in the system
headers there.


git-svn-id: https://svn.freebsd.org/base/stable/12@359471 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f